test(testConfig): 테스트용 헬퍼 분리
