+++ src/main/java/roomescape/config/ClockConfig.java
@@ -4,8 +4,10 @@ import java.time.Clock;
 import java.time.ZoneId;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 @Configuration
+@Profile("!test")
 public class ClockConfig {
 
     @Bean
