+++ src/test/java/roomescape/web/FourthMissionStepTest.java
@@ -5,11 +5,10 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import java.lang.reflect.Field;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.controller.ReservationController;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringWebTest
 public class FourthMissionStepTest {
 
     @Autowired
