+++ src/test/java/roomescape/web/SecondMissionStepTest.java
@@ -6,10 +6,9 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringWebTest
 public class SecondMissionStepTest {
 
     @Autowired
